mangohud: disable vsync
It always feels as though mangohud somehow interferes and causes screen tearing on the TV, no matter the value of mangohud's fps limit, and no matter of the vsync setting in-game.

It appears that explicitly disabling vsync in mangohud resolves this screen tearing issue on the TV. Vsync is then enabled in-game.