home-manager(we): fix wrong version of MPV being used
The mpv version was not the one created by home-manager, instead it used
the default one from pkgs. The 4k upscaling only worked because it
happened to be present in the `~/.config/mpv` directory. All other mpv
scripts/plugins are not present in that user config directory, hence did
seemingly not take any effect.