feat(fish)!: Use comma instead of custom fish_command_not_found
If a command was not found, the current behaviour was to simply try and
run the command using a flake from nixpkgs, using a very dumb:

nix run nixpkgs#$argv[1] -- $argv[2..]

This is not always desired in case the command was simply mistyped,
leading to an unecessary and annoying lookup to nixpkgs; especially in
case of typos. However, the ability to fetch unavailable commands from
nixpkgs is still quite useful, which is the use-case that comma tackles.

BREAKING CHANGE: fish_command_not_found does not fetch unavailable
command from nixpkgs anymore. Prepend `,` instead.