feat: Add pay-respects
pay-respects is out-of-the-way and does not highjack the commandline in
case of typos, as was previously the case with the custom
fish_command_not_found function that unconditionally tried to fetch the
command from nixpkgs.

pay-respects can be used for quick suggestions and fixes on the wrong
command, which also proves to be way more useful than simply fetching a
command from nixpkgs.