feat(we|mpv): Ability to disable anime4k profile
This is important in case the anime4k profile has not been enabled from
home-manager. In this case, `we` will make `mpv` fail to start because
it cannot find the anime4k profile in its configuration, which is used
for upscaling.

The MPV profile name has been abstracted into
`config.mpv.anime4kProfileName`. Furthermore, `we` now reads the MPV
profile name from an environment variable, where MPV's `--profile`
argument will not be set if the environment variable has not been
provided.

Lastly, we are letting home-manager take care of adding the profile to
MPV and using it in `we` in case anime4k has been enabled.