feat(swww): Do not assert path existence by default
This will require the derivation to be impure because path needs to be
checked which are outside of home-manager's scope. However, it may be
annoying to force an impure derivation every time for this little
check. If the path does not exist, the service will exit and the
wallpapers will simply not be changed periodically.

This new option allows to enable or not the assertions, in order to
require an impure evaluation with added checks to ensure the wallpaper
folder exists, or to skip this check and trust the user that path exists.